requirements.txt
pandas~=2.2.3
datasets~=3.6.0
fastapi~=0.115.12
scikit-learn~=1.6.1
python-dotenv~=1.1.0
torch~=2.6.0
transformers~=4.51.3
peft~=0.13.2
tqdm~=4.67.1
matplotlib~=3.10.3
uvicorn~=0.34.2
bitsandbytes~=0.45.2
chromadb~=0.4.24
nlpaug~=1.1.11
onnxruntime~=1.14.1